- flake8 and mypy - mypy rightly led us to modify the class hierarchy
