+++ run_configs/lfric/gungho.py
@@ -18,7 +18,6 @@ from fab.steps.grab import GrabFolder
 from fab.steps.link import LinkExe
 from fab.steps.preprocess import fortran_preprocessor
 from fab.steps.psyclone import psyclone_preprocessor, Psyclone
-from fab.util import run_command
 
 from grab_lfric import lfric_source_config, gpl_utils_source_config
 from lfric_common import Configurator, FparserWorkaround_StopConcatenation
