+++ system-tests-new/FortranDependencies/test_FortranDependencies.py
@@ -10,7 +10,6 @@ from fab.parse.fortran.fortran import AnalysedFortran
 
 from fab.build_config import BuildConfig
 from fab.constants import EXECUTABLES
-from fab.parse import AnalysedFile
 from fab.steps.analyse import Analyse
 from fab.steps.compile_c import CompileC
 from fab.steps.compile_fortran import CompileFortran
