+++ tests/unit_tests/steps/test_archive_objects.py
@@ -1,4 +1,3 @@
-from pathlib import Path
 from unittest import mock
 from unittest.mock import call
